Use daemon thread to report lull operation and move code to FnApiWorkerStatusHandler
